Polished game
Fix al of the mayor problems except 1 that has mini sections. When more
than one enemy is in warning range, the hud just goes invisible and if
they hit, the screen hit fx doesn't run for all of them, only the 1st to
hit. If the player dies when 2 hit while this bug is active, the meteors
keep running in the background. It has something to do with
isGameRuunung not being set to false. I'm not sure why the  hud goes
invisible.